+++ config/src/main/java/org/hyperledger/besu/config/StubGenesisConfigOptions.java
@@ -129,8 +129,8 @@ public class StubGenesisConfigOptions implements GenesisConfigOptions, Cloneable
   }
 
   @Override
-  public CliqueConfigOptions getCliqueConfigOptions() {
-    return CliqueConfigOptions.DEFAULT;
+  public JsonCliqueConfigOptions getCliqueConfigOptions() {
+    return JsonCliqueConfigOptions.DEFAULT;
   }
 
   @Override